requirements_windows.txt
cffi==1.15.0
cryptography==36.0.1
bitarray==1.5.3
cycler==0.10.0
kiwisolver==1.1.0
pycodestyle==2.5.0
pyparsing==2.4.6
pycparser==2.21
pyOpenSSL==21.0.0
python-dateutil==2.8.1
six==1.14.0
